% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ggseg_label.R
\name{add_ggseg_label}
\alias{add_ggseg_label}
\title{Add new column with ggseg ROI labels}
\usage{
add_ggseg_label(.data, feature_roi_col = "feature_roi", label_col = "label")
}
\arguments{
\item{.data}{larger merged dataframe}

\item{feature_roi_col}{name of existing column containing FreeSurfer-formatted ROI names from the DK atlas}

\item{label_col}{name of new ggseg label column (defaults to "label")}
}
\value{
dataframe with new column
}
\description{
This function transforms FreeSurfer-formatted ROI names from the
DK atlas into labels that are compatible with the \code{ggseg} R package for plotting
data onto a brain surface parcellation
}
