% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_add_suffix.R
\name{colname_add_suffix}
\alias{colname_add_suffix}
\title{Add suffix to column names}
\usage{
colname_add_suffix(.data, suffix, exclude = NULL)
}
\arguments{
\item{.data}{dataframe or tibble}

\item{suffix}{suffix to append to column names}

\item{exclude}{character vector containing columns to exclude; if not specified, then
all columns will be appended by default}
}
\value{
dataframe or tibble with new column names
}
\description{
This function appends a suffix to all columns of a dataframe or tibble.
One may additionally exclude columns from being appended by specifying them in the
\code{exclude} argument
}
\section{Notes}{

\itemize{
\item see https://stackoverflow.com/questions/52966616/exclude-some-columns-from-adding-a-suffix-in-r
for more details
}
}

