% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.match_data_slow}
\alias{.match_data_slow}
\title{Match data by date of entry (slow version)}
\usage{
.match_data_slow(
  ref_df,
  ref_col,
  match_col,
  select_col,
  group_col,
  max_diff = Inf,
  match_df = NULL,
  idx_col = NULL,
  diff_col = NULL,
  suffix = ".match"
)
}
\arguments{
\item{ref_col}{name of column to serve as reference value to match to}

\item{match_col}{name of column to get values to match}

\item{select_col}{names of column(s) to pull additional data from matched rows}

\item{group_col}{name of group column, usually a subject identifier column}

\item{max_diff}{(optional) the maximum allowed time difference for a valid match;
if no entires fall within this allowed time difference, NA will be matched}

\item{match_df}{(optional) a separate dataframe or tibble from which matched data
will be pulled from. If specified, then \code{match_col} and \code{select_col} will be
pulled from \code{match_df}. If not specified, then the aforementioned columns
are pulled from \code{.data}}

\item{idx_col}{(optional) string of name of column to store index from matched
dataframe}

\item{diff_col}{(optional) string of name of column to store absolute difference}

\item{suffix}{(optional) suffix appended to matched columns. If not specified,
defaults to ".match"}

\item{.data}{dataframe containing the reference data and matching data}
}
\value{
.data with additional columns containing matched data
}
\description{
Match a recorded data point to another recorded data point that is closest
in time to it
}
\details{
For example, match the CDR value closest in time to each subject's
PET scan. Each row containing non-NA data in the reference column will be
matched with a row containins non-NA data in the matching column which is closest
in value to the reference value. User may also specify additional columns to pull
data from the matched rows.
}
\section{Notes}{

When using dplyr::group_by, all subsequent functions on a dataframe
are performed per group. For our purposes, since we want to find the row index
corresponding to the matched entry for each reference value, the returned
index is of the group dataframe rather than the entire dataframe, which is not
the desired result. To combat this, we make a new column which indicates the
global index; that way, our matching function can reference this column and
return the global index rather than the group-wise index
}

\section{WARNING}{

This is an alternate implementation of \code{match_data}, but now using tidyverse
functions. The code itself is much more elegant, but in practice this function
takes more than 10x the amount of time to execute compared to the original.
Therefore I would recommend to just use the original, as it does exactly the
same thing as this implementation.
}

