% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_progressor.R
\name{get_progressor}
\alias{get_progressor}
\title{Mark subjects as stable or progressors based on chronological ordering of binary events}
\usage{
get_progressor(
  .data,
  event_col,
  time_col,
  group_col,
  event_label = NULL,
  baseline_col = NULL,
  time_window = Inf,
  stable_min_time = -Inf,
  progressor_nonevent_start = TRUE,
  exclude_revert = TRUE,
  exclude_stable_event = TRUE,
  keep_extra_col = FALSE,
  progressor_colname = "progressor"
)
}
\arguments{
\item{.data}{dataframe or tibble containing subject data to categorize as stable/progressor}

\item{event_col}{name of column containing event markings. This can either be a boolean column
where events are marked as TRUE, or a character column with multiple event/non-event labels. In
the latter case, specify argument \code{event_label} to indicate which of the labels correspond to a
true event}

\item{time_col}{name of column containing time information}

\item{group_col}{name of grouping column, usually a subject identifier column}

\item{event_label}{(optional) character vector containing the labels that correspond to a positive
event; specify this if \code{event_col} is not a boolean column}

\item{baseline_col}{(optional) name of column containing locations of the baseline to consider; if
not specified, defaults to baseline = subject's first entry in time}

\item{time_window}{(optional) length of time window from baseline to search for an event to check
for progressor status (default = Inf, i.e. look indefinitely into future)}

\item{stable_min_time}{(optional) length of minimum time that a stable subject must remain stable
to be labeled as such (default = -Inf, i.e. no minimum)}

\item{progressor_nonevent_start}{(optional) whether to impose constraint that progressors must start
as stable at baseline (default = TRUE)}

\item{exclude_revert}{(optional) whether to exclude reverters and categorize them as NA (default = TRUE)}

\item{exclude_stable_event}{(optional) whether to categorize a stable as NA if they have a positive
event at any time, even past the stable_min_time window}

\item{keep_extra_col}{(optional) whether to keep extra cols \code{.event}, which is the boolean event col
(useful when \code{event_col} is a character col), \code{.revert}, which indicates which subjects are reverters,
\code{.bl}, which indicates the baseline row, and \code{.time_to_progression}, which is the time from baseline to
first event}

\item{progressor_colname}{(optional) name of new column containing progressor labels (default =
"progressor")}
}
\description{
This function creates a new logical column named \code{progressor} which contains
subject-wise indications of progressor status. Subjects can fall into one of 3 categories:
\itemize{
\item progressor: subjects who start at event = FALSE and record an event = TRUE at a later time;
marked as TRUE
\item stable: subjects who start at event = FALSE and never record an event = TRUE; marked as FALSE
\item NA: subjects who are uncategorizable (see details)
}
}
\details{
\itemize{
\item Uncategorizable subjects can be due to one of three cases: (1) if a stable subject hasn't
remained stable for at least \code{stable_min_time}; (2) if a progressor subject doesn't start as
stable at baseline. These cases are marked as NA in the \code{progressor} column; (3) if a subject
progresses to an event but then at a later time reverts to the non-event (this is done only if
\code{exclude_revert = TRUE})
\item If a subject has more than one baseline specified in the baseline column, then the function defaults
to the first baseline in time
\item extra columns that can be kept via the \code{keep_extra_col} argument:
\itemize{
\item .event = logical column indicating locations of events (useful when \code{event_col} is a character col)
\item .revert = logical column indicating subject-wise reverter status
\item .bl = logical column indicating the baseline row
\item .time_to_progression = numeric column indicating time from baseline to first event
}
}
}
