% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subj_summary.R
\name{get_subj_summary}
\alias{get_subj_summary}
\title{Get subject-specific dataframe from a larger merged dataframe}
\usage{
get_subj_summary(.data, column_names, group_name = "Subject")
}
\arguments{
\item{.data}{larger merged dataframe}

\item{column_names}{vector of columns names to extract}

\item{group_name}{name of column to group entries by (usually subject ID)}
}
\value{
dataframe on the subject level with specified columns selected
}
\description{
Return dataframe where each row is a single subject, and each columns is a
subject-level field (e.g. date of first AV45+ scan, or first CDR record)
}
\details{
The purpose of this function is to return a dataframe where each row is on
the subject level, rather than on the entry level (as is the case with the
main ADNIMERGE tables). Since all rows will have the same value within-subject, the first row
of each subject is used to populate the subject-specific dataframe
}
