% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv_list.R
\name{load_csv_dir}
\alias{load_csv_dir}
\title{Load multiple CSVs in a single directory}
\usage{
load_csv_dir(d, vector = FALSE, bind = FALSE)
}
\arguments{
\item{d}{name of directory containing CSVs}

\item{vector}{if TRUE, then load CSVs without header; this is the case when the
CSVs only store a single vector of data}

\item{bind}{if TRUE, then load CSVs and bind into one dataframe, otherwise store
each individual dataframe in a list}
}
\value{
either list of dataframes or a single dataframe
}
\description{
This function loads multiple CSVs from a specified directory and
stores them in either a list (bind = FALSE) or a single dataframe (bind = TRUE)
}
