% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dir.R
\name{make_dir}
\alias{make_dir}
\title{Recursively make directory if doesn't exist}
\usage{
make_dir(d, verbose = FALSE)
}
\arguments{
\item{d}{path to directory}

\item{verbose}{if TRUE, then print status to terminal}
}
\description{
This function first checks to see whether a directory exists. If
it doesn't, it creates the directory and any necessary subdirectories. If it
does exist, it does nothing
}
