% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_event_censor.R
\name{mark_event_censor}
\alias{mark_event_censor}
\title{Create new column indicating events/right censor}
\usage{
mark_event_censor(
  .data,
  time_col,
  event_col,
  group_col,
  consecutive_events = 1
)
}
\arguments{
\item{.data}{dataframe with event data}

\item{time_col}{double column containing times of each entry (dplyr data-mask)}

\item{event_col}{logical column containing events, where TRUE indicates the
presence of an event (dplyr data-mask)}

\item{group_col}{name of grouping column, usually a subject ID column (string)}

\item{consecutive_events}{(optional) number of consecutive events that must occur
in order for it to be marked as a true event; if not specified, defaults to 1}
}
\value{
.data with additional columns containing event-censor markings
}
\description{
This function takes in 3 input columns: time, event and grouping.
Within each group, the function marks whether each entry is a non-event (0) or
event (1). In the case that no event exists for a particular group, the last
entry in time is marked as the right censor (2)
}
