% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_first_event.R
\name{mark_first_event}
\alias{mark_first_event}
\title{Mark where a partcular event occured first within each group}
\usage{
mark_first_event(.data, date_col, group_col, new_col)
}
\arguments{
\item{.data}{larger merged dataframe}

\item{date_col}{column containing event dates}

\item{group_col}{column containing group identifiers (usually subject IDs)}

\item{new_col}{name of new column which indicates first events}
}
\value{
dataframe with new column
}
\description{
This function creates a new boolean column (named with argument
\code{new_col}) which indicates where a particular event occured first within a
single group (usually a subject)
}
\details{
For this function to operate, one column must contain date or time of
event information (ex. date of a particular AV45 PET scan), and another column
must contain unique identifiers of groups (ex. subject IDs). These two are
inputted into the function as arguments \code{date_col} and \code{group_col}, respectively.
For each group ID, there will be only one TRUE entry which corresponds to the
earliest date in column \code{date_col}
}
