% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste0_na.R
\name{paste0_na}
\alias{paste0_na}
\title{Concatenate strings while propagating NAs}
\usage{
paste0_na(s_vec, s_paste)
}
\arguments{
\item{s_vec}{character vector}

\item{s_paste}{string to paste}
}
\value{
pasted character vector with NAs propagated
}
\description{
This function acts exactly like "paste0", only it propagates
NAs. This is useful for vectors which contain NAs, but the desired output
is NA rather than "NA{suffix}"
}
