% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_difference.R
\name{pivot_difference}
\alias{pivot_difference}
\title{Compute difference in long-formatted values using pivot_wider}
\usage{
pivot_difference(.data, names_col, values_col, ref, id_col = NULL)
}
\arguments{
\item{.data}{dataframe with data to compute difference}

\item{names_col}{name of col containing names of variable groupings}

\item{values_col}{name(s) of col containing values; can accept multiple value columns}

\item{ref}{name of the variable group to serve as the reference; must exist within the
\code{values_col} column}

\item{id_col}{(optional) vector of column names to serve as ID cols for pivot_wider}
}
\value{
dataframe with differences in values
}
\description{
This function uses a series of tidyr pivotting operations to compute
the difference in values compared to a reference name. This requires that the df
containing the data is in long format, and that the names of the variable types
are stored in a separate column.
}
\details{
First, the function runs \code{tidyr::pivot_wider} to convert the dataframe
into wide format. Then it computes the difference between each widened column and
the reference column. Finally it runs \code{tidyr::pivot_longer} to convert the dataframe
back into long format. Differences in values are stored in the original values column,
which by default has its values replaced by the differences.
}
