% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggseg.R
\name{plot_ggseg}
\alias{plot_ggseg}
\title{Plot data on a brain surface or coronal slice using ggseg}
\usage{
plot_ggseg(.data, fill_col, aseg_bool = FALSE, aseg_side = NULL)
}
\arguments{
\item{.data}{dataframe or tibble containing data to plot}

\item{fill_col}{name of column containing continuous-valued data}

\item{aseg_bool}{if true, plot on coronal slice instead of brain surface;
useful for visualizing subcortical ROIs}

\item{aseg_side}{string to indicate the view to plot for aseg if you only
want to plot one; either 'coronal' or 'sagittal'}
}
\value{
ggseg plot
}
\description{
This function plots continuous data on a brain surface or coronal
slice using the package \code{ggseg}
}
\details{
This function assumes that the input dataframe has a column named
"label" which contains ggseg-specific ROI labels of the FreeSurfer DK atlas
}
