% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggseg_patchwork.R
\name{plot_ggseg_patchwork}
\alias{plot_ggseg_patchwork}
\title{Plot multiple ggseg plots}
\usage{
plot_ggseg_patchwork(
  .data,
  fill_col,
  group_col,
  scale,
  cb_title = NULL,
  aseg_side = NULL
)
}
\arguments{
\item{.data}{dataframe or tibble containing data to plot on surface + aseg}

\item{fill_col}{string of column containing fill data}

\item{group_col}{string of column containing the group labels; this \strong{must} be a factor
column}

\item{scale}{ggplot2 fill scale; this will be applied to \emph{every} plot generated}

\item{cb_title}{(optional) title of colorbar; if none specified, defaults to \code{fill_col}}

\item{aseg_side}{string to indicate the view to plot for aseg if you only
want to plot one; either 'coronal' or 'sagittal'}
}
\value{
list of ggseg plots, with names in the form of "{group_level}.{surf|aseg}"
}
\description{
This function plots multiple brain surface + aseg ggseg plots, one per
group, then stores them in a list. This is useful for combining multiple plots into
one plot via patchwork's \code{wrap_plots} function.
}
