% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spaghetti.R
\name{plot_spaghetti}
\alias{plot_spaghetti}
\title{Create spaghetti plot for visualizing longitudinal subject-level data}
\usage{
plot_spaghetti(
  .data,
  col_x,
  col_group,
  col_color = NULL,
  col_shape = NULL,
  col_order = NULL,
  dodge = NULL
)
}
\arguments{
\item{.data}{dataframe or tibble containing longitudinal data}

\item{col_x}{column containing time values; this will typically be a
longitudinal metric, such as age or days from enrollment}

\item{col_group}{column containing unique subject identifiers}

\item{col_color}{(optional) column containing values to encode point colors}

\item{col_shape}{(optional) column containins values to encode point shapes}

\item{col_order}{(optional) column used to determine the ordering of the
y-axis index. For example, one may choose to order each "spaghetti string"
by date of AV45-positivity rather than age at enrollment. If not specified,
y-axis indices will be ordered by the minimum \code{col_x} value by default}
}
\value{
ggplot spaghetti plot
}
\description{
This function plots a spaghetti plot to visualize longitudinal data,
where the x-axis is the time axis and the y-axis is a discrete subject index
}
\details{
This function requires a dataframe in long-ish format where each row
corresponds to a record at a single time point for a single subject. The table
at minimum must contain a subject ID column and a time column. Values of other
features (e.g. intracranial volume, PET SUVR, amyloid-positivity, etc.) are
encoded via the shape (for discrete values) and color (for continuous values)
aesthetics. Additionally, one may order the y-axis indices based on \code{col_order},
which contains subject-level values (i.e. a single value corresponding to each
subject)
}
