% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surf_aseg.R
\name{plot_surf_aseg}
\alias{plot_surf_aseg}
\title{Plot ggseg surface + aseg plot}
\usage{
plot_surf_aseg(
  .data,
  fill_col,
  title_str = NULL,
  cb_title = NULL,
  aseg_side = NULL
)
}
\arguments{
\item{.data}{dataframe or tibble containing data to plot on surface + aseg}

\item{fill_col}{string of column containing fill data}

\item{title_str}{(optional) title of plot}

\item{cb_title}{(optional) title of colorbar; if none specified, defaults to \code{fill_col}}

\item{aseg_side}{string to indicate the view to plot for aseg if you only
want to plot one; either 'coronal' or 'sagittal'}
}
\value{
patchwork plot object
}
\description{
This function plots both a brain surface plot and brain aseg (volumetric)
plot using ggseg, then combines them using patchwork. The colorbars are collected
between the two plots, so the user should only specify one fill scale for the output
patchwork plot
}
