% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns in a dataframe}
\usage{
rename_columns(.data, col_old, col_new)
}
\arguments{
\item{.data}{dataframe with columns to rename}

\item{col_old}{vector of old column names}

\item{col_new}{vector of new column names; must be the same
length as col_old}
}
\value{
dataframe with renamed columns
}
\description{
This function renames column given a vector of old column names and
corresponding vector of new column names
}
