% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_and_rename_fs.R
\name{select_and_rename_fs}
\alias{select_and_rename_fs}
\title{Select and rename FreeSurfer ROI-based features from ADNI or OASIS merged table}
\usage{
select_and_rename_fs(
  .data,
  col_select,
  study_str = "adni",
  feature_type_str = NULL
)
}
\arguments{
\item{.data}{ADNI or OASIS merged dataframe}

\item{col_select}{character vector containing additional columns to select}

\item{study_str}{indicates study, either "adni" or "oasis" (must be lowercase);
defaults to adni}

\item{feature_type_str}{(optional) indicates feature type, either "suvr" or "vol";
if not specified, then both feature types will be selected}
}
\value{
dataframe with selected and renamed columns
}
\description{
This function first selects columns corresponding to FreeSurfer
ROI-based features from the ADNI or OASIS datasets, along with other user-
specified columns, then renames the FS columns to FreeSurfer standard ROI
labels as notated in the color lookup table
}
